Rename the generator project
